Refactor BitcoinExchange methods to use double for date value calculations, improving precision. Update getDateValue and calculBtc methods to handle double values instead of int, ensuring accurate results when processing Bitcoin data.

+++ Module09/ex00/BitcoinExchange.hpp
@@ -22,7 +22,7 @@ class BitcoinExchange {
 
 		bool	isValidExtension(std::string fileName);
 
-		int		getDateValue(std::string date);
+		double		getDateValue(std::string date);
 
 		void	calculBtc();
 		void	printInput();
